+++ src/common/hooks/useRecordControls.ts
@@ -5,10 +5,11 @@ import { postRecord, putRecord, deleteRecord as deleteRecordRequest } from '@com
 import { PROFILE_BFIDS } from '@common/constants/bibframe.constants';
 import { StatusType } from '@common/constants/status.constants';
 import { DEFAULT_RECORD_ID } from '@common/constants/storage.constants';
-import { deleteRecordLocally, formatRecord, getRecordId } from '@common/helpers/record.helper';
+import { deleteRecordLocally, getRecordId } from '@common/helpers/record.helper';
 import { UserNotificationFactory } from '@common/services/userNotification';
 import { useConfig } from '@common/hooks/useConfig.hook';
 import { getSavedRecord } from '@common/helpers/record.helper';
+import { formatRecord } from '@common/helpers/recordFormatting.helper';
 import { getRecord } from '@common/api/records.api';
 import { ROUTES } from '@common/constants/routes.constants';
 import state from '@state';
