+++ src/test/__tests__/common/helpers/record.helper.test.ts
@@ -3,7 +3,6 @@ import * as ProgressBackupHelper from '@common/helpers/progressBackup.helper';
 import { localStorageService } from '@common/services/storage';
 import { AUTOCLEAR_TIMEOUT } from '@common/constants/storage.constants';
 import * as BibframeConstants from '@src/common/constants/bibframe.constants';
-import * as BibframeMappingConstants from '@common/constants/bibframeMapping.constants';
 import { getMockedImportedConstant } from '@src/test/__mocks__/common/constants/constants.mock';
 
 describe('record.helper', () => {
@@ -21,178 +20,11 @@ describe('record.helper', () => {
     data: record,
   };
   const testInstanceUri = 'testInstanceUri';
-  const testInstantiatesUri = 'testInstantiatesUri';
-  const testInstantiatesToInstanceUri = 'testInstantiatesToInstanceUri';
   const mockTypeUriConstant = getMockedImportedConstant(BibframeConstants, 'TYPE_URIS');
-  const mockInstantiatesToInstanceConstant = getMockedImportedConstant(
-    BibframeConstants,
-    'INSTANTIATES_TO_INSTANCE_FIELDS',
-  );
-  const mockBFLiteUriConstant = getMockedImportedConstant(BibframeMappingConstants, 'BFLITE_URIS');
-  const mockNonBFRecordElementsConstant = getMockedImportedConstant(BibframeMappingConstants, 'NON_BF_RECORD_ELEMENTS');
   mockTypeUriConstant({ INSTANCE: testInstanceUri });
-  mockInstantiatesToInstanceConstant([]);
-  mockBFLiteUriConstant({ INSTANTIATES: testInstantiatesUri, NOTE: 'testNoteUri' });
-  mockNonBFRecordElementsConstant({ testNoteUri: { container: '_notes' } });
 
   beforeEach(() => {
     jest.spyOn(ProgressBackupHelper, 'generateRecordBackupKey').mockReturnValue(key);
-  });
-
-  describe('formatRecord', () => {
-    function testFormatRecord(initialRecord: ParsedRecord, testResult: Record<string, object | string>) {
-      const result = RecordHelper.formatRecord(initialRecord);
-
-      expect(result).toEqual(testResult);
-    }
-
-    test('returns formatted record data', () => {
-      const initialRecord = {
-        [testInstanceUri]: {},
-      };
-      const testResult = {
-        resource: {
-          [testInstanceUri]: {},
-        },
-      };
-
-      testFormatRecord(initialRecord, testResult);
-    });
-
-    test('embeds work entity into instance if there is data for work entity', () => {
-      const workComponent = {
-        testUri: 'testValue',
-      };
-      const initialRecord = {
-        [testInstanceUri]: {},
-        [testInstantiatesUri]: workComponent,
-      };
-      const testResult = {
-        resource: {
-          [testInstanceUri]: {
-            [testInstantiatesUri]: [workComponent],
-          },
-        },
-      };
-
-      testFormatRecord(initialRecord, testResult);
-    });
-
-    test("doesn't embed work entity into instance if it's empty", () => {
-      const workComponent = {};
-      const initialRecord = {
-        [testInstanceUri]: {},
-        [testInstantiatesUri]: workComponent,
-      };
-      const testResult = {
-        resource: {
-          [testInstanceUri]: workComponent,
-        },
-      };
-
-      testFormatRecord(initialRecord, testResult);
-    });
-  });
-
-  describe('updateInstantiatesWithInstanceFields', () => {
-    test('returns initial Instance object', () => {
-      const instance = {
-        [testInstantiatesUri]: [{}],
-      } as unknown as Record<string, RecursiveRecordSchema>;
-
-      const result = RecordHelper.updateInstantiatesWithInstanceFields(instance);
-
-      expect(result).toEqual(instance);
-    });
-
-    test('returns updated Instance object that contained Instantiates', () => {
-      mockInstantiatesToInstanceConstant([testInstantiatesToInstanceUri]);
-      const instance = {
-        [testInstantiatesUri]: [{ existingKey_1: ['existingUri_1'] }],
-        [testInstantiatesToInstanceUri]: ['testUri_1', 'testUri_2'],
-      } as unknown as Record<string, RecursiveRecordSchema[]>;
-      const testResult = {
-        [testInstantiatesUri]: [
-          {
-            existingKey_1: ['existingUri_1'],
-            [testInstantiatesToInstanceUri]: ['testUri_1', 'testUri_2'],
-          },
-        ],
-      };
-
-      const result = RecordHelper.updateInstantiatesWithInstanceFields(instance);
-
-      expect(result).toEqual(testResult);
-    });
-
-    test('returns initial Instance object that did not contain Instantiates', () => {
-      mockInstantiatesToInstanceConstant([testInstantiatesToInstanceUri]);
-      const instance = {
-        [testInstantiatesToInstanceUri]: ['testUri_1', 'testUri_2'],
-      } as unknown as Record<string, RecursiveRecordSchema[]>;
-      const testResult = {
-        [testInstantiatesUri]: [
-          {
-            [testInstantiatesToInstanceUri]: ['testUri_1', 'testUri_2'],
-          },
-        ],
-      };
-
-      const result = RecordHelper.updateInstantiatesWithInstanceFields(instance);
-
-      expect(result).toEqual(testResult);
-    });
-  });
-
-  describe('updateRecordWithDefaultNoteType', () => {
-    test('returns initial record', () => {
-      const record = {
-        _notes: [
-          {
-            value: ['note description 1'],
-            type: ['testType_1'],
-          },
-          {
-            value: ['note description 2'],
-            type: ['testType_2'],
-          },
-        ],
-      };
-
-      const result = RecordHelper.updateRecordWithDefaultNoteType(record);
-
-      expect(result).toEqual(record);
-    });
-
-    test('returns a record with a default Note type', () => {
-      const record = {
-        _notes: [
-          {
-            value: ['note description 1'],
-            type: ['testType_1'],
-          },
-          {
-            value: ['note description 2'],
-          },
-        ],
-      } as Record<string, RecursiveRecordSchema[]>;
-      const testResult = {
-        _notes: [
-          {
-            value: ['note description 1'],
-            type: ['testType_1'],
-          },
-          {
-            value: ['note description 2'],
-            type: ['testNoteUri'],
-          },
-        ],
-      };
-
-      const result = RecordHelper.updateRecordWithDefaultNoteType(record);
-
-      expect(result).toEqual(testResult);
-    });
   });
 
   test('deleteRecordLocally - invokes "localStorageService.delete" with generated key', () => {
